% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngmr-package.R
\docType{package}
\name{ngmr-package}
\alias{ngmr}
\alias{ngmr-package}
\title{ngmr: Ngram Markov Model}
\description{
Fit and evaluate ngram Markov models.
}
\details{
ngmr: Fit and evaluate ngram Markov models.
}
\author{
\strong{Maintainer}: Joonas Miettinen \email{joonas.m.miettinen@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
